chore: Deprecating old models and updating them to zod
